// ClientApp/src/app/app-routing.enum.ts
export enum AppRoutingEnum {
  Home = '',
  PersonalData = 'donnees_personnelles',
  Contact = 'contact',
  Rendezvous = 'reservation',
  Help = 'aide',
  MyRendezvous = 'mes_rendezvous',
  EditProfil = 'modificationProfil',
  PlanningPersonnel = 'planningPersonnel',
  ChoosePortal = 'choisir-portail',
  PortalPersonnelSignIn = 'portail-personnel-signUp',
  PortalFamilleSignIn = 'portail-famille-signUp',
  PortalPersonnelLogin = 'portail-personnel-login',
  PortalFamilleLogin = 'portail-famille-login',
  ListContacts = "list-contacts",
  ResidentPersonnel = 'resident',
  DonneesPerso = 'donneesPersonnelles',
  MentionsLegales = 'mentionsLegales'
}